backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-environ
psycopg2-binary
gunicorn
dotenv